update_Events/requirements.txt
sqlalchemy
python-dotenv
kafka-python
psycopg2-binary
pymysql